GPII-1987: Splitted authDataLoader.js into 2 js files that one contains the actual implementation of the auth loader component and the other is to call the loader component to perform the loading so that the latter can be enhanced to accept an configuration.
